mac compilation correction

+++ src/wallet/rpcwallet.cpp
@@ -1827,7 +1827,7 @@ static void ListTransactions(CWallet* const pwallet, const CWalletTx& wtx, const
             {
                 WalletTxToJSON(wtx, entry);
                 size_t dataSize=wtx.tx->GetOP_ReturnSize();
-                entry.pushKV("datasize", dataSize);
+                entry.pushKV("datasize", static_cast<int>(dataSize));
             }
             entry.pushKV("abandoned", wtx.isAbandoned());
             ret.push_back(entry);
